changed timestamp to nanoseconds, added presence

+++ mediapipe/java/com/google/mediapipe/solutions/lindera/ComputerVisionPlugin.java
@@ -1,5 +1,5 @@
 package com.google.mediapipe.solutions.lindera;
 
 public interface ComputerVisionPlugin {
-    void bodyJoints(int timestamp, BodyJoints bodyJoints);
+    void bodyJoints(long timestamp, BodyJoints bodyJoints);
 }
